requirements.txt
django>=4.2,<5.1
mysqlclient>=2.2
requests>=2.31
python-dotenv>=1.0
